# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Crepe & Waffle API documentation'
  description: 'A documented API for the required requests.'
  contact:
    email: ahmed.bltagy@appssquare.com
  version: 1.0.0
servers:
  -
    url: 'https://app.crepe-waffle.com/api'
    description: 'Production Server'
paths:
...
  /SendContactUs:
    post:
      tags:
        - 'Contact Us'
      summary: 'Send the contact us form'
      description: ''
      operationId: contactUs
      parameters:
        -
          name: id
          in: query
          description: 'User Id in case logged in'
          required: false
          schema:
            type: string
        -
          name: name
          in: query
          description: 'full name'
          required: false
          schema:
            type: string
        -
          name: email
          in: query
          description: 'User email'
          required: false
          schema:
            type: string
        -
          name: photo
          in: query
          description: 'attached photo base64'
          schema:
            type: string
        -
          name: type
          in: query
          description: 'The type of message ( suggestion, complain or comment)'
          schema:
            type: string
        -
          name: message
          in: query
          description: 'The form message'
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'The User object with token to use it on SMS verification message'
        422:
          description: 'Validation Error'
      security:
        -
          petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /homeSlider:
    get:
      tags:
        - 'Sliders&gallery'
      summary: 'Retrieve home sliders for website and app'
      description: ''
      operationId: homeSlider
      responses:
        200:
          description: ''
        422:
          description: 'Validation Error'
      security:
        -
          petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /serviceRate:
    post:
      tags:
        - Orders
      summary: 'Submit a rate for a service'
      description: 'Authentication required on header using token'
      operationId: serviceRate
      parameters:
        -
          name: user_id
          in: query
          description: 'User ID ( already retrieved )'
          required: true
          schema:
            type: integer
        -
          name: service_rate
          in: query
          description: "the object that have the rate it should look like\n\n    {'crepe':5,'waffle ':5,'delivery':5} "
          required: true
          schema:
            type: string
        -
          name: order_id
          in: query
          description: 'The Order id that you got from submitting order'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: ''
        422:
          description: 'Validation Error'
      security:
        -
          petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
components:
  securitySchemes:
    'Password Based':
      type: oauth2
      description: 'Use a global client_id / client_secret and your username / password combo to obtain a token'
      in: header
      flows:
        password:
          authorizationUrl: /api/login
          tokenUrl: /api/login
          scopes: {  }
      scheme: https
...